Move clock helpers to covenant-utils
