+++ contracts/clock/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod contract;
 pub mod error;
-pub mod helpers;
 pub mod msg;
 pub mod state;
 pub mod test_helpers;
