+++ contracts/interchain-router/Cargo.toml
@@ -21,7 +21,6 @@ library = []
 
 [dependencies]
 covenant-macros  = { workspace = true }
-valence-clock   = { workspace = true, features = ["library"] }
 cosmwasm-schema  = { workspace = true }
 cosmwasm-std     = { workspace = true }
 cw-storage-plus  = { workspace = true }
