+++ contracts/stride-liquid-staker/src/contract.rs
@@ -5,14 +5,16 @@ use cosmwasm_std::{
     to_json_binary, to_json_string, Binary, Deps, DepsMut, Env, MessageInfo, Reply, Response,
     StdError, StdResult, Uint128,
 };
-use covenant_utils::ica::{
-    get_ica, msg_with_sudo_callback, prepare_sudo_payload, query_ica_registration_fee, sudo_error,
-    sudo_open_ack, sudo_response, sudo_timeout, INTERCHAIN_ACCOUNT_ID,
-};
 use covenant_utils::neutron::{self, get_proto_coin, RemoteChainInfo, SudoPayload};
+use covenant_utils::{
+    clock::{enqueue_msg, verify_clock},
+    ica::{
+        get_ica, msg_with_sudo_callback, prepare_sudo_payload, query_ica_registration_fee,
+        sudo_error, sudo_open_ack, sudo_response, sudo_timeout, INTERCHAIN_ACCOUNT_ID,
+    },
+};
 use cw2::set_contract_version;
 use neutron_sdk::query::min_ibc_fee::MinIbcFeeResponse;
-use valence_clock::helpers::{enqueue_msg, verify_clock};
 
 use crate::helpers::{Autopilot, AutopilotConfig};
 use crate::msg::{ContractState, ExecuteMsg, InstantiateMsg, MigrateMsg, QueryMsg};
