+++ contracts/swap-holder/Cargo.toml
@@ -21,7 +21,6 @@ cosmwasm-schema  = { workspace = true }
 cosmwasm-std     = { workspace = true }
 cw-storage-plus  = { workspace = true }
 cw-utils         = { workspace = true }
-valence-clock   = { workspace = true, features = ["library"] }
 cw2              = { workspace = true }
 serde            = { workspace = true }
 thiserror        = { workspace = true }
