+++ contracts/swap-holder/src/msg.rs
@@ -4,10 +4,10 @@ use cosmwasm_std::{
 };
 use covenant_macros::{clocked, covenant_clock_address, covenant_deposit_address};
 use covenant_utils::{
-    instantiate2_helper::Instantiate2HelperConfig, CovenantPartiesConfig, CovenantTerms,
+    clock::dequeue_msg, instantiate2_helper::Instantiate2HelperConfig, CovenantPartiesConfig,
+    CovenantTerms,
 };
 use cw_utils::Expiration;
-use valence_clock::helpers::dequeue_msg;
 
 use crate::state::CONTRACT_STATE;
