+++ contracts/two-party-pol-holder/Cargo.toml
@@ -21,7 +21,6 @@ covenant-macros = { workspace = true }
 cosmwasm-schema = { workspace = true }
 cosmwasm-std    = { workspace = true }
 cw-storage-plus = { workspace = true }
-valence-clock  = { workspace = true, features = ["library"] }
 cw2             = { workspace = true }
 thiserror       = { workspace = true }
 schemars        = { workspace = true }
